+++ app/src/test/java/org/akvo/caddisfly/util/ColorUtilsTest.java
@@ -1,8 +1,15 @@
 package org.akvo.caddisfly.util;
 
 import android.graphics.Color;
+import android.util.Pair;
 
 import junit.framework.TestCase;
+
+import org.akvo.caddisfly.model.ResultRange;
+import org.akvo.caddisfly.model.TestInfo;
+
+import java.util.Hashtable;
+import java.util.List;
 
 public class ColorUtilsTest extends TestCase {
 
@@ -12,7 +19,14 @@ public class ColorUtilsTest extends TestCase {
     }
 
 //    public void testGetPpmValue() throws Exception {
-//        Bitmap bitmap = Bitmap.createBitmap(50, 50, Bitmap.Config.ARGB_8888);
+//        int[] colors = new int[2500];
+//        for (int i = 0; i < 50; i++) {
+//            for (int j = 0; j < 50; j++) {
+//                colors[i] = -1;
+//            }
+//        }
+//
+//        Bitmap bitmap = Bitmap.createBitmap(colors, 5, 5, Bitmap.Config.ARGB_8888);
 //        ArrayList<ResultRange> arrayList = new ArrayList<>();
 //        Bundle bundle = ColorUtils.getPpmValue(bitmap, arrayList, 50);
 //        assertEquals(null, bundle);
@@ -29,7 +43,19 @@ public class ColorUtilsTest extends TestCase {
     }
 
     public void testAutoGenerateColors() throws Exception {
+        Hashtable hashtable = new Hashtable();
+        TestInfo testInfo = new TestInfo(hashtable, "FLUOR", "ppm", 0);
 
+        for (int i = 0; i < 5; i++) {
+            ResultRange resultRange = new ResultRange(((int) (Double.valueOf(i) * 10)) / 10f, Color.TRANSPARENT);
+            testInfo.addRange(resultRange);
+        }
+
+        List list = ColorUtils.autoGenerateColors(testInfo);
+        for (int i = 0; i < list.size(); i++) {
+            assertEquals(String.format("FLUOR-%.2f", i * 0.1), ((Pair) list.get(i)).first);
+            assertEquals(-16777216, ((Pair) list.get(i)).second);
+        }
     }
 
     public void testGetColorFromRgb() throws Exception {